% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitter.R
\name{comment_command}
\alias{comment_command}
\title{Prepare a command to be an argument}
\usage{
comment_command(command, quote = "'")
}
\arguments{
\item{command}{(\code{character}) The command to convert.}
}
\value{
(\code{character})
}
\description{
Prepare a command to be an argument
}

