% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitter.R
\name{make_submit_script}
\alias{make_submit_script}
\title{Make qsub submission script}
\usage{
make_submit_script(command, parallel = FALSE, out_file = NULL,
  err_file = NULL, name = NULL)
}
\arguments{
\item{command}{(\code{character}) One or more commands to run}

\item{parallel}{(\code{logical} of length 1) If \code{TRUE}, run jobs in parallel.}

\item{out_file}{(\code{character} of length 1) Where job runtime output is stored.}

\item{err_file}{(\code{character} of length 1) Where job error output is stored.}

\item{name}{(\code{character} of length 1) The job name.}
}
\value{
(\code{character} of length 1) The text for a qsub submission script.
}
\description{
Makes the text for a qsub submission script.
}

