% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitter.R
\name{qstat}
\alias{qstat}
\title{Return parsed qstat table}
\usage{
qstat(remote, quiet = FALSE)
}
\arguments{
\item{remote}{(\code{\link{remote_server}}) The remote server information.}

\item{quiet}{(\code{logical} of length 1) Supress messeges.
Default: \code{FALSE}}
}
\value{
\code{data.frame}
}
\description{
Execute \code{qstat} and parse the qstat table output
}
\examples{
\dontrun{

remote <- remote_server$new(server = "shell.somewhere.edu", user = "joeshmo", port = 345)
qstat(remote)
}
}
