% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitter.R
\name{qsub}
\alias{qsub}
\title{Submit jobs to qsub over ssh}
\usage{
qsub(command, remote, parallel = FALSE, remote_cwd = NULL,
  runtime_folder = NULL)
}
\arguments{
\item{command}{(\code{character}) One or more commands to run}

\item{remote}{(\code{\link{remote_server}}) The remote server information.}

\item{parallel}{(\code{logical} of length 1) If \code{TRUE}, run jobs in parallel.
Default: \code{FALSE}}

\item{remote_cwd}{(\code{character} of length 1)
What to set the current working directory to on the remote server before running jobs.
Default: The home directory on the remote server.}

\item{runtime_folder}{(\code{character} of length 1)
The location to save the runtime error, output, and qsub submission script.
Default: A unique name with timestamp, random string, and first command in job.}
}
\description{
This function submits commands to qsub (SGE) over ssh.
It assumes password-less ssh keys are set up.
}
\examples{
\dontrun{

remote <- remote_server$new(server = "shell.somewhere.edu", user = "joeshmo", port = 345)
qsub(c('echo "test 1" > test1.txt', 'echo "test 2" > test2.txt'), remote, runtime_folder = "~/test/out")
}
}

