% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitter.R
\name{qsubmit}
\alias{qsubmit}
\title{Submit jobs to qsub over ssh}
\usage{
qsubmit(command, remote, parallel = FALSE, runtime_folder = NULL)
}
\arguments{
\item{command}{(\code{character}) One or more commands to run}

\item{remote}{(\code{\link{remote}})}

\item{parallel}{(\code{logical} of length 1)}
}
\description{
This function submits commands to qsub (SGE) over ssh.
It assumes password-less ssh keys are set up.
}

