% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitter.R
\name{random_char}
\alias{random_char}
\title{Random character}
\usage{
random_char(count = 5)
}
\arguments{
\item{count}{(\code{numeric} of length 1) The length of the random string.}
}
\description{
Make a random character of a specified length
}
