% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitter.R
\name{remote_r}
\alias{remote_r}
\title{Run R commands on remote server}
\usage{
remote_r(remote, variables = list(), expression, ...)
}
\arguments{
\item{remote}{(\code{\link{remote_server}}) The remote server information.}

\item{variables}{(\code{list}) Variables that the \code{expression} relies on.
Their content will be deparsed and transfered to the remote server.}

\item{expression}{An expression to run on a remote server.}

\item{...}{Passed to \code{\link{ssh_command}}}
}
\description{
Run R commands on remote server.
Variables in the local envrioment can be transfered to the remote computer.
}
\examples{
\dontrun{
remote <- remote_server$new(server = "shell.somewhere.edu", user = "joeshmo", port = 345)
a = 1:10
f = function(x) {
  x = x + 2
  x*2
}
remote_r(remote, list(a, f), {
  a = a + 3
  f(a)
})
}
}

