% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitter.R
\name{ssh_command}
\alias{ssh_command}
\title{Run a commmand on a remote server}
\usage{
ssh_command(command, remote, quote = "'")
}
\arguments{
\item{command}{(\code{character}) The command to run}

\item{remote}{(\code{\link{remote}})}
}
\description{
Run a commmand on a remote server using ssh.
}

