% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitter.R
\name{wait_for_qsub}
\alias{wait_for_qsub}
\title{Wait for qsub job}
\usage{
wait_for_qsub(remote, job_name, quiet = TRUE)
}
\arguments{
\item{remote}{(\code{\link{remote_server}}) The remote server information.}

\item{job_name}{(\code{character} of length 1) The name of the job.}

\item{quiet}{(\code{logical} of length 1) Supress messeges.
Default: TRUE}
}
\description{
Wait for qsub job to complete or fail.
}

