% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitter.R
\name{qsub}
\alias{qsub}
\title{Submit jobs to qsub over ssh}
\usage{
qsub(command, remote, parallel = FALSE, remote_cwd = NULL,
  runtime_folder = "qsub_records", wait = TRUE, cores = 1,
  print_track = FALSE, print_info = TRUE, echo = TRUE, stdout = FALSE,
  stderr = FALSE)
}
\arguments{
\item{command}{(\code{character}) One or more commands to run}

\item{remote}{(\code{\link{remote_server}}) The remote server information.}

\item{parallel}{(\code{logical} of length 1) If \code{TRUE}, run jobs in parallel.
Default: \code{FALSE}}

\item{remote_cwd}{(\code{character} of length 1)
What to set the current working directory to on the remote server before running jobs.
Default: The home directory on the remote server.}

\item{runtime_folder}{(\code{character} of length 1)
The location to save the runtime error, output, and qsub submission script.
Default: A unique name with timestamp, random string, and first command in job.}

\item{wait}{(\code{logical} of length 1) If \code{TRUE}, the function will wait for all
jobs to complete before exiting.
Default: TRUE}

\item{cores}{(\code{numberic}) The number of cores to use.
Default: 1}

\item{print_track}{(\code{logical} of length 1) If \code{TRUE}, the progress of the jobs is periodically printed.
Default: TRUE}

\item{print_info}{(\code{logical} of length 1) If \code{TRUE}, information on the job is printed
Default: TRUE}

\item{echo}{(\code{logical} of length 1) If \code{TRUE}, the command to be executed is printed
Default: TRUE}

\item{stdout}{(\code{logical} or \code{character} of length 1) If \code{TRUE}, print or save the standard output of the job.}

\item{stderr}{(\code{logical} or \code{character} of length 1) If \code{TRUE}, print or save the standard error of the job.}
}
\description{
This function submits commands to qsub (SGE) over ssh.
It assumes password-less ssh keys are set up.
}
\examples{
\dontrun{

remote <- remote_server$new(server = "shell.somewhere.edu", user = "joeshmo", port = 345)
qsub(c('echo "test 1" > test1.txt', 'echo "test 2" > test2.txt'), remote, runtime_folder = "~/test/out")
}
}

