% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitter.R
\docType{class}
\name{remote_server}
\alias{remote_server}
\title{Class for remote server}
\format{\code{\link{R6Class}} object.}
\usage{
remote_server
}
\value{
Object of \code{\link{R6Class}} with methods for communication with lightning-viz server.
}
\description{
Stores information needed to connect to a remote user
}
\section{Fields}{

\describe{
\item{\code{server}}{Address of the server}

\item{\code{user}}{User name}

\item{\code{port}}{Port of remote server}
}}
\section{Methods}{

\describe{
  \item{new}{Make new instance of this class}
  }
}
\keyword{datasets}

