% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitter.R
\name{rsync_push}
\alias{rsync_push}
\title{Rsync files to remote server.}
\usage{
rsync_push(local_path, remote_path, remote)
}
\arguments{
\item{remote}{(\code{\link{remote_server}}) The remote server information.}

\item{file_paths}{(\code{character})
Paths of files on the local machine for transfer.}
}
\description{
Copys files to a remote server using rsync.
}

